% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{influx.connection}
\alias{influx.connection}
\title{Create a connection object to an InfluxDB}
\usage{
influx.connection(host = "127.0.0.1", port = 8086, scheme = "http",
  user = "", pass = "")
}
\arguments{
\item{host}{hostname}

\item{port}{numerical. port number}

\item{scheme}{NA}

\item{user}{username}

\item{pass}{password}
}
\description{
Create a connection object to an InfluxDB
}
