% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{influx.query}
\alias{influx.query}
\title{Query an influxdb server}
\usage{
influx.query(queries, con, db = "tickdb")
}
\arguments{
\item{queries}{The Influxdb queries to send}

\item{con}{An influx_connection object}

\item{db}{Sets the name of the database}
}
\description{
Query an influxdb server
}
