% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{load_idb}
\alias{load_idb}
\title{Load the instrument database}
\usage{
load_idb(config)
}
\arguments{
\item{config}{file name of the database json config file}
}
\description{
Load the instrument database
}
