% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{period}
\alias{period}
\title{Helpers function to make a data.frame to specify a time period like 08:00,17:00 with 2 integers only}
\usage{
period(from, to)
}
\arguments{
\item{from}{start hour}

\item{to}{end hour}
}
\description{
Helpers function to make a data.frame to specify a time period like 08:00,17:00 with 2 integers only
}
