% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{plot.sampleprice}
\alias{plot.sampleprice}
\title{Plot a sampleprice object}
\usage{
\method{plot}{sampleprice}(x, ...)
}
\arguments{
\item{x}{an object of class sampleprice, usually, a result of a call to 'sample_price'}
}
\description{
Plot a sampleprice object
}

