% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{plot.tickdata}
\alias{plot.tickdata}
\title{Plot a tickdata object}
\usage{
\method{plot}{tickdata}(x, ...)
}
\arguments{
\item{x}{an object of class tickdata, usually, a result of a call to 'tick_data'}
}
\description{
Plot a tickdata object
}

