% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{print.sampleprice}
\alias{print.sampleprice}
\title{Print a sampleprice object}
\usage{
\method{print}{sampleprice}(x, ...)
}
\arguments{
\item{x}{an object of class sampleprice, usually, a result of a call to 'sample_price'}
}
\description{
Print a sampleprice object
}
