% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{print.tickdata}
\alias{print.tickdata}
\title{Print a tickdata object}
\usage{
\method{print}{tickdata}(x, ...)
}
\arguments{
\item{x}{an object of class tickdata, usually, a result of a call to 'tick_data'}
}
\description{
Print a tickdata object
}
