% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{raw_sample}
\alias{raw_sample}
\title{Run a simple query with a group by argument}
\usage{
raw_sample(measurement, product, type, expiry, from, to, group, config)
}
\arguments{
\item{measurement}{book,trade}

\item{product}{Product name}

\item{type}{Product type}

\item{from}{UTC nanosecond}

\item{to}{UTC nanosecond}

\item{group}{influx group by clause}

\item{config}{json config file}
}
\description{
Run a simple query with a group by argument
}
