% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{run.query}
\alias{run.query}
\title{Run a TickDB query}
\usage{
run.query(query, db, sample = F, stype = "", con)
}
\arguments{
\item{Influx}{DB connection (use influx.connection)}

\item{an}{InfluxDB query for the Tick database}
}
\value{
a data.frame
}
\description{
Run a TickDB query
}
