% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{sample_price}
\alias{sample_price}
\title{sample price series}
\usage{
sample_price(measurement, product, type, from, to, periods, frequency = "1s",
  front = 1, rolldays = 5,
  config = "~/recherche/tickdatabase/influx/config.json")
}
\description{
sample price series
}
