% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{seq.contracts}
\alias{seq.contracts}
\title{generate sequence of contract and business date for a product, and the nanoseconds timestamps of required periods}
\usage{
\method{seq}{contracts}(product, type, front, rolldays, from, to, periods, idb)
}
\arguments{
\item{product}{the product's name like FDAX or ODAX}

\item{type}{the product's type like F,O,E,S or C}

\item{front}{1 for front month, 2 for back month, 3 for 3rd month, etc...}

\item{rolldays}{the number of days to roll the contract before its expiry date}

\item{from}{the start date of the period}

\item{to}{the end date of the period}

\item{periods}{a data.frame with 2 columns, from and to, of trading time period like "08:00" "16:30"}

\item{idb}{the instrument database or the json config file used to load the database}
}
\description{
generate sequence of contract and business date for a product, and the nanoseconds timestamps of required periods
}
