% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{summary.sampleprice}
\alias{summary.sampleprice}
\title{Produce result summaries on tick data}
\usage{
\method{summary}{sampleprice}(x, ...)
}
\arguments{
\item{x}{an object of class sampleprice, usually, a result of a call to 'sample_price'}
}
\description{
Produce result summaries on tick data
}

