% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{summary.tickdata}
\alias{summary.tickdata}
\title{Produce result summaries on tick data}
\usage{
\method{summary}{tickdata}(x, ...)
}
\arguments{
\item{x}{an object of class tickdata, usually, a result of a call to 'tick_data'}
}
\description{
Produce result summaries on tick data
}

