% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{tick_data}
\alias{tick_data}
\title{Sample price series}
\usage{
tick_data(measurement, product, type, from, to, periods, front = 1,
  rolldays = 5, fields = NULL,
  config = "~/recherche/tickdatabase/influx/config.json", verbose = F)
}
\arguments{
\item{measurement}{l1book,l2book,l2.5book,trade}

\item{product}{Product name}

\item{type}{Product type}

\item{from}{start date}

\item{to}{end date}

\item{periods}{trading hours}

\item{front}{1 for front month, 2 for back month, etc...}

\item{rolldays}{days to roll the contract before expiry}

\item{config}{json config file}

\item{frequency}{sampling frequency}
}
\description{
Sample price series
}

