% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb-packages.R
\docType{package}
\name{tickdb-package}
\alias{tickdb-package}
\title{tickdb}
\description{
Functions to access the Liquid Capital Tick Database
}
\author{
David Bellot \email{dbellot@liquidcapital.com}
}

