% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{make.query}
\alias{make.query}
\title{Generate a query from specifications}
\usage{
make.query(db, con, measurement, product, type, from, to, periods, front,
  rolldays, fields, extended_result = F)
}
\description{
Generate a query from specifications
}
