% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{period}
\alias{period}
\title{Make a 'periods' data.frame like 08:00,17:00 with 2 integers only}
\usage{
period(from, to)
}
\arguments{
\item{from}{start hour}

\item{to}{end hour}
}
\description{
Make a 'periods' data.frame like 08:00,17:00 with 2 integers only
}
