% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickdb.r
\name{run.query}
\alias{run.query}
\title{Run a predefined query and return a list of data.frame with the results}
\usage{
run.query(query, con = NULL, db = "tickdb", sample = F, stype = "")
}
\arguments{
\item{Influx}{DB connection (use influx.connection)}

\item{an}{InfluxDB query for the Tick database
export run.query}
}
\description{
Run a predefined query and return a list of data.frame with the results
}
